Implemented sampler

+++ raytracer/samplers/Simple.hpp
@@ -9,6 +9,7 @@
    Courtesy Kevin Suffern.
 */
 
+#include <stddef.h>
 #include <vector>
 #include "Sampler.hpp"
 
@@ -32,7 +33,7 @@ class Simple : public Sampler {
   virtual ~Simple() = default;
 
   // Shoot a ray of weight 1 through the center of the pixel.
-  std::vector<Ray> get_rays(int px, int py) const;
+  std::vector<Ray> get_rays(size_t px, size_t py) const;
 };
 
 #endif  // RAYTRACER_SAMPLERS_SIMPLE_HPP_
